Fix Azure authentication failing when day-of-month < 10
[Re #1]
We were using the RFC-1123 to build the "x-ms-date" header that is
used for Azure authentication, but apparently it's not correct. The
RFC uses 1 or 2 digits for the day of month representation, but MS
expects the date in the "HTTP Date format", which apparently always
uses 2 digits for the day of month.