Fix get-object sometimes returning object as text
[Re #2]
When objects had the content-type set to something that the http-kit
library interprets as text, the result 'object' was being returned as
a 'string' instead of a 'input-stream.

The fix involves asking http-kit to always return an
'input-stream'. If the response is succesfull we don't need to do
anything more. If there is an error, we need to read the input-stream
and return it as text (for consistency with other operations).

As we are specifying a specific API version, we can assume MS will
keep a consistent behaviour and always return errors in XML.

An alternative approach would be to keep the http-kit automatic
content detection, and to convert strings to input-streams when
required. But it would be trickier to avoid memory issues because we
have no control over the size of the objects. In the case of error
messages we can assume they won't be (too) big.